requirements.txt
gradio
Pillow
